scripts/archive-grabber/requirements.txt
internetarchive>=3.0.0
PyYAML>=6.0
mutagen>=1.47.0
